`pytest-beartype 0.3.0` pre-flight x 4.
This commit is the next in a commit chain preparing for our upcoming
`pytest-beartype 0.3.0` release. Specifically, this commit significantly
refactors the implementation of this plugin for sanity, readability, and
maintainability. Of course, doing so inadvertently broke this plugin...
which, itself, is insane. For safety, this breakage is currently
isolated to this branch. In short, it's bad. (*Badly busted, buster! Sadly flustered!*)