Synchronous generator fixtures x 5.
This commit is the next in a commit chain generalizing the
`--beartype-fixtures` option accepted by this plugin to transparently
type-check **synchronous generator fixtures** (i.e., fixtures whose
signatures are *not* prefixed by the `async` keyword and whose bodies
contain one or more `yield` statements), en-route to publishing our
upcoming `pytest-beartype 0.3.0` release. Specifically, this commit:

* Refactors test-specific paths to be encapsulated by memoized
  `pathlib.Path` objects in the same manner as the main `beartype_test`
  test suite.
* Correctly configures the `pytest` configuration file copied into the
  `pytest` subprocesses forked by the `test_option_beartype_tests()`
  and `test_option_beartype_fixtures()` integration tests.

Now we only have to test everything and hope against hope for worky!
It's not much. Face is exhausted. Yet the urge to code remains strong.
(*Impressive impression on a massive missive!*)